+++ docs/reference-guides/core-blocks.md
@@ -539,15 +539,6 @@ Display a post's featured image. ([Source](https://github.com/WordPress/gutenber
 -	**Supports:** align (center, full, left, right, wide), color (~~background~~, ~~text~~), spacing (margin, padding), ~~html~~
 -	**Attributes:** height, isLink, scale, sizeSlug, width
 
-## Post Link
-
-Add the link of this post.
-
--	**Name:** core/post-link
--	**Category:** theme
--	**Supports:** color (background, gradients, ~~text~~), spacing (margin, padding), typography (fontSize, lineHeight), ~~html~~
--	**Attributes:** content, linkTarget
-
 ## Post Navigation Link
 
 Displays the next or previous post link that is adjacent to the current post. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/post-navigation-link))
@@ -664,6 +655,15 @@ Give quoted text visual emphasis. "In quoting others, we cite ourselves." — Ju
 -	**Category:** text
 -	**Supports:** anchor, typography (fontSize, lineHeight)
 -	**Attributes:** align, citation, value
+
+## Read More
+
+Displays the link of a post, page, or any other content-type. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/read-more))
+
+-	**Name:** core/read-more
+-	**Category:** theme
+-	**Supports:** color (background, gradients, ~~text~~), spacing (margin, padding), typography (fontSize, lineHeight), ~~html~~
+-	**Attributes:** content, linkTarget
 
 ## RSS
