+++ lib/blocks.php
@@ -93,7 +93,6 @@ function gutenberg_reregister_core_block_types() {
 				'post-date.php'                    => 'core/post-date',
 				'post-excerpt.php'                 => 'core/post-excerpt',
 				'post-featured-image.php'          => 'core/post-featured-image',
-				'post-link.php'                    => 'core/post-link',
 				'post-navigation-link.php'         => 'core/post-navigation-link',
 				'post-terms.php'                   => 'core/post-terms',
 				'post-title.php'                   => 'core/post-title',
@@ -104,6 +103,7 @@ function gutenberg_reregister_core_block_types() {
 				'query-pagination-numbers.php'     => 'core/query-pagination-numbers',
 				'query-pagination-previous.php'    => 'core/query-pagination-previous',
 				'query-title.php'                  => 'core/query-title',
+				'read-more.php'                    => 'core/read-more',
 				'rss.php'                          => 'core/rss',
 				'search.php'                       => 'core/search',
 				'shortcode.php'                    => 'core/shortcode',
