+++ packages/block-library/src/index.js
@@ -70,7 +70,6 @@ import * as postContent from './post-content';
 import * as postDate from './post-date';
 import * as postExcerpt from './post-excerpt';
 import * as postFeaturedImage from './post-featured-image';
-import * as postLink from './post-link';
 import * as postNavigationLink from './post-navigation-link';
 import * as postTemplate from './post-template';
 import * as postTerms from './post-terms';
@@ -85,6 +84,7 @@ import * as queryPaginationPrevious from './query-pagination-previous';
 import * as queryTitle from './query-title';
 import * as quote from './quote';
 import * as reusableBlock from './block';
+import * as readMore from './read-more';
 import * as rss from './rss';
 import * as search from './search';
 import * as separator from './separator';
@@ -195,7 +195,6 @@ export const __experimentalGetCoreBlocks = () => [
 	postAuthor,
 	postDate,
 	postTerms,
-	postLink,
 	postNavigationLink,
 	postTemplate,
 	queryPagination,
@@ -203,6 +202,7 @@ export const __experimentalGetCoreBlocks = () => [
 	queryPaginationNumbers,
 	queryPaginationPrevious,
 	postComments,
+	readMore,
 	logInOut,
 	termDescription,
 	queryTitle,
