+++ packages/block-library/src/read-more/style.scss
@@ -1,4 +1,4 @@
-.wp-block-post-link {
+.wp-block-read-more {
 	display: block;
 	width: fit-content;
 	&:not([style*="text-decoration"]) {
