+++ packages/block-library/src/style.scss
@@ -28,7 +28,6 @@
 @import "./post-comments-form/style.scss";
 @import "./post-excerpt/style.scss";
 @import "./post-featured-image/style.scss";
-@import "./post-link/style.scss";
 @import "./post-terms/style.scss";
 @import "./post-title/style.scss";
 @import "./preformatted/style.scss";
@@ -36,6 +35,7 @@
 @import "./post-template/style.scss";
 @import "./query-pagination/style.scss";
 @import "./quote/style.scss";
+@import "./read-more/style.scss";
 @import "./rss/style.scss";
 @import "./search/style.scss";
 @import "./separator/style.scss";
