--- test/integration/fixtures/blocks/core__post-link.html
@@ -1,1 +1,0 @@
-<!-- wp:post-link /-->
