--- test/integration/fixtures/blocks/core__post-link.serialized.html
@@ -1,1 +1,0 @@
-<!-- wp:post-link /-->
