+++ test/integration/fixtures/blocks/core__read-more.html
@@ -1,0 +1,1 @@
+<!-- wp:read-more /-->
