+++ test/integration/fixtures/blocks/core__read-more.parsed.json
@@ -1,6 +1,6 @@
 [
 	{
-		"blockName": "core/post-link",
+		"blockName": "core/read-more",
 		"attrs": {},
 		"innerBlocks": [],
 		"innerHTML": "",
